Refactor lints in clippy_lints::attrs into separate submodules/files
